+++ lib/grape.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'logger'
+require 'monitor'
 require 'active_support'
 require 'active_support/isolated_execution_state'
 require 'active_support/core_ext/array/conversions' # to_xml